ultimed firefox debug edition
debug by Nepha

- Girls, team,
I upgrade the mac with tools and firefox Edition. So we found that config is more efficient.

Please note we debug the flex and css grid is easier with FFD.

So, Neah please, finish the card. Nepha Code and review please.

Nephelim > MD the project for exam.

With my love Nepha